+++ server/symfony/src/Entity/Lookup.php
@@ -6,7 +6,7 @@ use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
-#[ORM\Entity]
+#[ORM\Entity(repositoryClass: "App\Repository\LookupRepository")]
 #[ORM\Table(name: 'lookups')]
 class Lookup
 {